Move file validation logic to a validator class
