+++ ComicBookRegistry.Domain/Validation/FileValidator.cs
@@ -1,0 +1,41 @@
+﻿using ComicBookRegistry.Domain.Constants;
+using ComicBookRegistry.Domain.Exceptions;
+using System.IO;
+using System.Linq;
+
+namespace ComicBookRegistry.Domain.Validation
+{
+    public class FileValidator
+    {
+        private readonly string[] _acceptedExtensions =
+        {
+            FileConstants.Jpg,
+            FileConstants.Jpeg,
+            FileConstants.Png,
+            FileConstants.Bmp,
+        };
+
+        public void Validate(FileInfo file)
+        {
+            if (file == null)
+            {
+                throw new NullFileException();
+            }
+
+            if (file.Length == 0)
+            {
+                throw new EmptyFileException();
+            }
+
+            if (file.Length > FileConstants.MaxFileBytes)
+            {
+                throw new MaximumFileSizeExceededException();
+            }
+
+            if (_acceptedExtensions.All(extension => extension != Path.GetExtension(file.Name).ToLower()))
+            {
+                throw new InvalidFileTypeException();
+            }
+        }
+    }
+}
